consolidate safeio test scaffolding
